msm: camera_v2: Add index info in camera SMMU handle
In the current design, we do not have any index inforamtion of
handle. So we need to traverse through the whole SMMU table to
get the corresponding index by comparing the index. Now we add
index info in the handle so camera buf manager can directly
use same index with camera SMMU driver. In addition,
provide number of SMMU clients to buf manager.

Change-Id: I7602e55df2974dc1c651e87c737b7791ac225637
Signed-off-by: Yiduo Wang <yiduow@codeaurora.org>