msm: cpp: Dequeue the enqueued frame from queue in case of error
When scheduling a frame fails, enqueued frame from processing
queue is not dequeued. Once the frame command is deleted as part
of error handling, accessing the processing queue list causes
invalid access. Dequeue the frame from the processing queue list
in case of error and reset the timer if there is no element in the
queue.

CRs-Fixed: 901551
Change-Id: Id676dbfa725014341f07c1dbdbdb97667e03dc9f
Signed-off-by: Krishnankutty Kolathappilly <kkolatha@codeaurora.org>